refactor: Extract vote buttons into separate reusable component
